MetaSync/cloud-functions/metasync/requirements.txt
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
functions-framework==3.4.0